Created icons and screenshots

+++ Basketball 2022/ViewController.swift
@@ -71,7 +71,7 @@ class ViewController: UIViewController, ARSCNViewDelegate {
         ballNode.physicsBody = SCNPhysicsBody(type: .dynamic, shape: SCNPhysicsShape(node: ballNode))
         
         // Calculate force for pushing the ball
-        let power = Float(5)
+        let power = Float(10)
         let x = -matrixCameraTransform.m31 * power
         let y = -matrixCameraTransform.m32 * power
         let z = -matrixCameraTransform.m33 * power
@@ -95,7 +95,7 @@ class ViewController: UIViewController, ARSCNViewDelegate {
             type: .static,
             shape: SCNPhysicsShape(node: hoopNode,
             options: [SCNPhysicsShape.Option.type: SCNPhysicsShape.ShapeType.concavePolyhedron]
-            )
+                )
         )
         
         return hoopNode
